+++ tests/testthat/test-trace.R
@@ -4,6 +4,10 @@ context("trace.R")
 test_that("tree printing only changes deliberately", {
   skip_unless_utf8()
 
+  # Because of srcrefs
+  skip_on_cran()
+  skip_if_not_installed("testthat", "2.99.0")
+
   local_options(
     rlang_trace_format_srcrefs = TRUE,
     `rlang:::trace_force_dangling_srcrefs` = TRUE
@@ -30,6 +34,10 @@ test_that("tree printing only changes deliberately", {
 
 test_that("can print tree with collapsed branches", {
   skip_unless_utf8()
+
+  # Because of srcrefs
+  skip_on_cran()
+  skip_if_not_installed("testthat", "2.99.0")
 
   # Fake eval() call does not have same signature on old R
   skip_if(getRversion() < "3.4")
@@ -203,7 +211,8 @@ test_that("eval() frames are collapsed", {
 
 test_that("%>% frames are collapsed", {
   skip_unless_utf8()
-  skip_if_not_installed("magrittr")
+  skip_if_not_installed("magrittr", "1.5.0.9000")
+  skip_on_cran()
 
   # Fake eval() call does not have same signature on old R
   skip_if(getRversion() < "3.4")
@@ -227,7 +236,8 @@ test_that("%>% frames are collapsed", {
 
 test_that("children of collapsed %>% frames have correct parent", {
   skip_unless_utf8()
-  skip_if_not_installed("magrittr")
+  skip_if_not_installed("magrittr", "1.5.0.9000")
+  skip_on_cran()
 
   # Fake eval() call does not have same signature on old R
   skip_if(getRversion() < "3.4")
@@ -288,7 +298,8 @@ test_that("pipe_collect_calls() collects calls", {
 
 test_that("combinations of incomplete and leading pipes collapse properly", {
   skip_unless_utf8()
-  skip_if_not_installed("magrittr")
+  skip_if_not_installed("magrittr", "1.5.0.9000")
+  skip_on_cran()
 
   # Fake eval() call does not have same signature on old R
   skip_if(getRversion() < "3.4")
@@ -323,7 +334,8 @@ test_that("combinations of incomplete and leading pipes collapse properly", {
 
 test_that("calls before and after pipe are preserved", {
   skip_unless_utf8()
-  skip_if_not_installed("magrittr")
+  skip_if_not_installed("magrittr", "1.5.0.9000")
+  skip_on_cran()
 
   # Fake eval() call does not have same signature on old R
   skip_if(getRversion() < "3.4")