+++ NEWS.md
@@ -1,6 +1,8 @@
 
 # rlang (development version)
 
+* Compatibility with upcoming testthat 3 and magrittr 2 releases.
+
 * `get_env()` now returns the proper environment with primitive
   functions, i.e. the base namespace rather than the base environment
   (r-lib/downlit#32).